face.storage/requirements.txt
alembic==1.12.1
Mako==1.2.4
MarkupSafe==2.1.3
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
